FIXED: HTTPS: Avoid leaking file descriptor after failed TLS negotation.
Two issues must be distinguished here:

1) ssl_exit/1 was a no-op here, because ssl_context/3 is used instead of
   the (now deprecated) ssl_init/3 to initialize the connection

2) if the TLS negotiation failed, then the exception was thrown but the
   plain stream was not closed.

A test case for this is to start the example server with:

    $ swipl -g server server.pl

and then run the following query from a different Prolog process:

    ?- repeat,
       catch(http_open('https://localhost:1111', Stream, []), _, true),
       false.

In this case, TLS negotiation always fails (because the certificate
cannot be verified; also, no client certificate is specified).

Without this patch, running the query leaks file descriptors.