ADDED: New option close_notify(+Boolean), making the server send        close_notify when closing the connection.
In addition, enabling this option raises an exception when EOF is
encountered without either side having initiated a TLS shutdown. This
mitigates truncation attacks, in which an adversary may cut the
connection or inject TCP FIN packets to simulate EOF.

This attack only applies to protocols that are *not* self-terminating
and is therefore extremely rare in practice (well-designed protocols
are self-terminating). However, the client/server examples are of this
type, and therefore this option is now enabled in the examples.