FIXED: Make client close the underlying plain streams in unit tests.
This omission was detected by the preceding improvement for handling
fatal alerts. In this situation, the server waits for EOF or TLS
shutdown, and because the client has previously received a timeout, it
does not send a shutdown. Therefore, all depends on properly closing
the underlying socket in this case.