ENHANCED: Use OpenSSL default curves (ecdh_curve/1) in OpenSSL >= 1.1.0.
Earlier OpenSSL versions do not set a curve by default. We use
prime256v1 in such cases, as before, but no longer override the new
library defaults with this curve.

This commit makes stronger curves automatically available when using
newer OpenSSL versions, unless users have explicitly asked for a
specific curve. A major attraction of OpenSSL 1.1.0 is X25519, which
is the new default curve in that version for ECDH.

To disable ECDH entirely, use the cipher_list/1 option (also as before).

In the future, ecdh_curve/1 will be superseded by a new option that
binds SSL_CTX_set1_groups. However, this currently affects both ECDH
and ECDSA, and we will adopt the eventual solution of OpenSSL.