MODIFIED: ssl_add_certificate_key/3 has become ssl_add_certificate_key/4.
This avoids destructive modifications to a context. It is important to
preserve the original context in these operations, because we cannot
_remove_ a certificate due to limitations of the current OpenSSL API.

An even more important reason is of course to preserve *logical purity*.
Note how logical purity automatically makes the code thread safe!

More about logical purity: https://www.metalevel.at/prolog/purity

In addition, I will provide new hooks that will allow you to replace a
certificate while the server is running. This is now possible because
new certificates can be added non-destructively, yielding a copy.

The advantage of this API over creating a context from scratch is that
the *password* of the original context is preserved. This means that
an attacker would have to gain access not only to Prolog, but actually
has to gain access to main memory in order to read the password.

This allows us to store the password in such a way that the web server
can no longer access it after dropping privileges.