ADDED: ssl_set_sni_hook/3, allowing to set the SNI callback of a context.
This makes it possible to later override a hook that is created for
example by the HTTP Unix daemon when the context is created.

This predicate lets you enjoy the convenience of the HTTP Unix daemon
together with the flexibility of custom code, among other benefits.

Note again how *logical purity* makes this predicate thread safe. This
is especially interesting in this case because not even OpenSSL itself
is thread safe for such use cases!