MODIFIED: Remove and ignore obsolete option cert/1.
cert(+Bool) controlled whether a certificate is sent. This was always
the *default* for servers, and *automatically* set to true if a
certificate and key were specified for a client.

This option was already documented as obsolete for several years.
Now, it has become even more meaningless since also the server can be
started without any certificate, and certificates and keys can be
added later with the new ssl_add_certificate_key/4 predicate. If you
do not want a certificate to be sent, do not supply one. If you supply
a certificate and key, it will be automatically used, exactly as
before.  This is *the same* for both server and client role.

An additional issue with this option was that its rather short and
non-descriptive name overlaps with the much more important option
peer_cert/1, which also takes a Boolean argument. This made the use of
this library unnecessarily harder and more error-prone.

It is still OK if the option is specified. It is simply ignored. Note
that sending the certificate (if it is supplied) can hardly have any
negative impact on security, and it is extremely unlikely that users
have specified a certificate but do not want it to be sent. Therefore,
the option was almost completely useless.