MODIFIED: library(crypto): Rename "pkcs" to "pkcs1" in options.
The padding options 'pkcs1' and 'pkcs1_oaep' have long been documented
in the manual, yet the code incorrectly used pkcs and pkcs_oaep instead.

This commit aligns the code with the documented and common name for
these padding options. The two affected options must be renamed if
they are currently used explicitly in any application.