MODIFIED: API change of rsa_sign/4 and rsa_verify/4, see below for more.
The semantics of rsa_sign/4 and rsa_verify/4 are now changed as follows:

  1) All *signatures* are now hex-encoded instead of raw bytes.
     This affects both rsa_sign/4 and rsa_verify/4. This enhances
     readability and exchange of signatures, and aligns the format
     with that of ecdsa_sign/4, ecdsa_verify/4 as well as that of the
     hash predicates and also other values that occur in this context.

  2) The default *encoding* of Data is now the new value 'hex'.
     This means that we assume that a hex-encoded string, atom or list
     is to be signed, and we automatically convert it to raw bytes.
     This is the most common case, since all hashes of library(crypto)
     are represented in this way, and we typically sign such digests.
     Again, this affects both rsa_sign/4 and rsa_verify/4.

  3) If the signature cannot be verified, rsa_verify/4 now *fails*,
     whereas it would previously throw an exception. This change makes
     the handling of failed verification a bit easier and more natural.

All code samples and uses within this package are adapted. Please review!

If you are using these predicates in your application, you need to
adapt them. In many cases, your code will become simpler now, if you
use the new hashing predicates crypto_data_hash/3 etc. of
library(crypto), since they all yield digests in hexadecimal notation.

If you have handled signatures as completely opaque objects, you will
likely only be affected by point (2). A (comparatively) typical case
is that you are currently using one of the other hashing libraries,
possibly convert the hash somehow, and then use rsa_sign/4 to sign the
result. In such cases, simply use crypto_data_hash/3 instead: It
yields a hash that can be used directly for signing with rsa_sign/4
(and also with ecdsa_sign/4) using the new semantics.

Alternatively, you can simply specify [encoding(octet)] to use the
encoding of Data that was previously the default. Note though that the
representation of the signature is still the new format in that case!

If [encoding(hex)] (i.e, the new default) is used, a *domain error* is
now thrown if Data is not in hexadecimal format. In most cases, this
lets you quickly detect whether your application needs to be adapted.

For example, the query in b380cad24cbe5102a549bfb8f7406e34a2b99e1a,
where rsa_verify/4 is used, can now be simplified to:

    ?- key(Key),
       to_be_signed(Hs),
       phrase(bytes_hex(Bs), Hs),
       crypto_data_hash(Bs, Digest, [algorithm(sha256),encoding(octet)]),
       signature(Sig),
       rsa_verify(Key, Digest, Sig, [type(sha256)]).

This is because the hexadecimal encoding that is already used in the
representation of certificates, keys and hashes can now be seamlessly
used in the RSA and ECDSA authentication predicates as well.