ADDED: hmac/1 option for all hash predicates of library(crypto).
A hash-based message authentication code (HMAC) can be used to
simultaneously verify the integrity *and* authenticity of data.

The new hmac(+Key) option is now available for all hash predicates
of library(crypto), enabling HMAC computation with a given key.

This option is a significant generalization of hmac_sha/4 from
library(sha): Starting with OpenSSL 1.1.0, hash-based MACs are now
available for *all* provided digest algorithms with library(crypto).

library(xmldsig) is a use case that needs this functionality.