remove a dubious workaround
What is the "synchronization issue in OpenSSL" that this is supposed
to correct? I have not found anything about this. If there still is
such an issue, it is better to bring it fully to light and to solve it
more elegantly. In fact, the workaround may cause a different problem:

      Therefore, in order to perform non-default initialisation,
      OPENSSL_init_ssl() MUST be called by application code prior to
      any other OpenSSL function calls.

Let us first and foremost ensure that we are following the advice of
the OpenSSL documentation.