ADDED: New predicate ssl_set_options/3 for tweaking SSL contexts.
This predicate is an important building block for setting various SSL
parameters from hooks, such as when using the HTTP Unix daemon.  Note
the important design principle: *No destructive modification to
existing contexts!* This is important to guarantee thread safety on
the Prolog level. Logical purity all the way.

Usage examples of this new predicate include:

  -) disabling specific protocol versions
  -) setting elliptic curves for key exchange
  -) enforcing client authentication via certificates
  -) etc.

For example, to disable SSLv3, you can add the following definition to
your server:

    http:ssl_server_create_hook(SSL0, SSL, _) :-
        ssl_set_options(SSL0, SSL, [disable_ssl_methods([sslv3])]).

This is of course only necessary for old versions of OpenSSL, since
SSLv3 is disabled in all versions that are still supported (>= 1.0.2).