MODIFIED: rsa_sign/4 and rsa_verify/4 now use sha256 by default.
Previously, sha1 was the default. I expect zero impact from this
change, because every programmer who uses these predicates *must*
think about the actual hash algorithm that ought to be used on signing
or verification. It has for several years been known that SHA-1 cannot
be considered secure, and in fact even a concrete collision was
recently found. Thus, any existing application must have been migrated
to a different algorithm due to security considerations, making the
old default inapplicable. Also the internal uses of these predicates
all specify the actual algorithm as an option to these predicates, and
hence no further changes were needed anywhere within SWI-Prolog due to
this modification.

The primary purpose of this change is to signal that security is being
taken seriously here, and to do everything that is possible to prevent
future users from relying on insecure algorithms (however unlikely
that may be, since most will explicitly specify a secure algorithm).

The new default of sha256 is also in alignment with what the hash
predicates such as crypto_data_hash/3 produce by default.