MODIFIED: The default hash algorithms are subject to change. See below.
The design principle and guarantee is this: The default algorithm used
for hashing is _cryptographically secure_. To allow transitioning to
secure default algorithms also in the future, you must not rely on the
particular algorithm that is used by default. Currently, the default
is sha256. In the future, a different algorithm may become the
default. If you specify an explicit algorithm, then that algorithm is
of course used.

All hash predicates, and rsa_verify/4 and rsa_sign/4 are affected.

To make interoperability and compatibility possible, the affected
predicates *all* yield the actual algorithm that was used if the
pertaining option is present and the algorithm is a logical variable.