CLEANUP: "context" --> "hash_context" in internal definitions.
There are different types of "contexts" in OpenSSL. We need to
distinguish hash contexts from cipher contexts in preparation of
incremental encryption, which will become available in SWI 7.8.

The public library(crypto) API is not affected by this change.

There are 3 public predicates for incremental hashing:

    crypto_context_new/2
    crypto_data_context/3
    crypto_context_hash/2

There may be slightly better names, and I will reconsider the API of
library(crypto) once incremental encryption is also available.  It
should be rather easy to smoothly transition the hash API because
these predicates are rarely used. crypto_file_hash/3 is more important
for hashes of large files, and that name is perfectly OK.